More models

+++ src/main/kotlin/cga/exercise/components/camera/TronCamera.kt
@@ -12,8 +12,8 @@ import shader.ShaderProgram
  */
 
 class TronCamera(
-    var fieldofview :Float= Math.toRadians(90F),
-    var aspectratio:Float = 16F/9F,
+    var fieldofview:Float= Math.toRadians(90F),
+    var aspectratio: Double = 16F/9F,
     var nearplane:Float = 0.1F,
     var farplane:Float = 100F) : ITron, Transformable() {
