+++ src/main/kotlin/cga/exercise/components/geometry/Transformable.kt
@@ -2,7 +2,6 @@ package cga.exercise.components.geometry
 
 import org.joml.Matrix4f
 import org.joml.Vector3f
-import java.beans.beancontext.BeanContextChild
 
 open class Transformable(private var modelMatrix: Matrix4f = Matrix4f(), var parent: Transformable? = null) {
     /**
@@ -31,7 +30,7 @@ open class Transformable(private var modelMatrix: Matrix4f = Matrix4f(), var par
      * @param yaw radiant angle around y-axis ccw
      * @param roll radiant angle around z-axis ccw
      */
-    fun rotate(pitch: Float, yaw: Float, roll: Float) {
+    fun rotate(pitch: Float, yaw: Double, roll: Float) {
         modelMatrix.rotateXYZ(pitch, yaw, roll)
     }
